requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==40.0.2
dnacentersdk==2.6.11
fastjsonschema==2.19.1
Flask==2.2.3
future==0.18.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.24.4
pandas==1.5.3
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.2
zipp==3.17.0